Moved BufferWriter to utils folder
